+++ tests/JsonApi/Posts/Hydrator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CloudCreativity\LaravelJsonApi\Tests\JsonApi\Posts;
+
+use CloudCreativity\LaravelJsonApi\Hydrator\EloquentHydrator;
+
+class Hydrator extends EloquentHydrator
+{
+}